Update randomFood.py
starting the loop, work on the if statements at home. The switch cases might be hard. Try setting the restaurants in an array? Print the resterauint in the array that is the number randomly picked -1? remove the food place from the list if it has been picked and rejected? move it to another list?